fix collab rights

+++ resources/views/notes/index.blade.php
@@ -13,19 +13,26 @@
         </div>
         <div class="d-flex align-items-center">
         <span data-filter-by="text">{{$note->created_at->diffForHumans()}}</span>
-        <div class="ml-1 dropdown card-options">
-            <button class="btn-options" type="button" id="note-dropdown-button-1" data-toggle="dropdown" aria-haspopup="true" aria-expanded="false">
-            <i class="material-icons">more_vert</i>
-            </button>
-            <div class="dropdown-menu dropdown-menu-right">
-                <a href="{{route('notes.edit', $note->id)}}" class="dropdown-item" data-remote="true">
-                    {{__('Edit')}}
-                </a>
-                <a href="{{route('notes.destroy', $note->id)}}" class="dropdown-item text-danger" data-method="delete" data-remote="true">
-                    {{__('Delete')}}
-                </a>
+        @can('update', $note)
+            <div class="ml-1 dropdown card-options">
+                <button class="btn-options" type="button" id="note-dropdown-button-1" data-toggle="dropdown" aria-haspopup="true" aria-expanded="false">
+                <i class="material-icons">more_vert</i>
+                </button>
+                <div class="dropdown-menu dropdown-menu-right">
+                    @can('update', $note)
+                        <a href="{{route('notes.edit', $note->id)}}" class="dropdown-item" data-remote="true">
+                            {{__('Edit')}}
+                        </a>
+                    @endcan
+                    @can('delete', $note)
+                        <div class="dropdown-divider"></div>
+                        <a href="{{route('notes.destroy', $note->id)}}" class="dropdown-item text-danger" data-method="delete" data-remote="true">
+                            {{__('Delete')}}
+                        </a>
+                    @endcan
+                </div>
             </div>
-        </div>
+        @endcan
         </div>
     </div>
     <div class="card-body p-1" data-filter-by="text">
